print/textlive-base: use a local base in configure
Replace /usr/local/ paths with local base paths in configure.

Instead, it would be better to regenerate configure with a replaced
prefix path. However, since there is a patch modifying configure in the
port, that could result in additional issues in the future and such
simple replacement might work better in practice.