sysutils/nut: Move patch file back to its proper place
I had moved the file to the root of port's directory to perform some
testing when updating the port to 2.8.0 and forgot to move it to its
rightful place in files/.

The patch file avoids redundantly building systemd and Solaris support
files that are never used in FreeBSD thereby reducing build time.
PORTREVISION doesn't need to change.

Reported by:	lwhsu
MFH:		No sense doing that so late in the quarter