sysutils/fusefs-ntfs: Support device offset to mount
Usually NTFS volume located on device at some offset.
We can use losetup/mdconfig for map device from offset to another device and mount this device.
But there are some problems:

It is new resourse, which has limited && can be leak
Method of mount is not portabled
Need copy data twice: for loop device and fuse
If we pass argument for device offset we can avoid such problems.