ports-mgmt/portmaster: Fix an issue with flavored dependencies
Flavored run dependencies of an unflavored port could cause the
installation of the port to fail after it had been successfully
built.

This issue should be fixed by this update, which executes a previously
commented out function call to update the FLAVOR in the environment of
the make command used to install the port.

I do not remember why I had commented out that function call, a long
time ago - but I do not expect it to cause any regressions, since the
same FLAVOR is now set enforced during "make install" as had been used
to build the port.

PR:		265084
Reported by:	alt2600 at icloud.com