databases/rrdtool: add new port option LOCKTIMEOUT
By default, RRDTool tries to lock RRD file and fails
if it cannot obtain the lock immediately.

New build option LOCKTIMEOUT (disabled by default)
allows to change this behavior and specify
a time interval (in seconds) with environment variable
RRDTOOL_LOCK_TIMEOUT to wait for lock if the file is busy.

It will fail if the lock cannot be obtained in time.
Zero value makes it wait for the lock indefinitely.

The patch documents the feature in the manual page and HTML docs.

Default build is not affected, so PORTREVISION not bumped.