Revert "devel/wasi-compiler-rt14: add new port"
Charlie Li <vishwin@freebsd.org> requests a revert of the addition of
this port, since it needs updates to wase-libc and wasi-libc++ to be
used.

The support for LLVM-14 has been worked on since May of 2022, with a
number of unresolved issues remaining (see reviews D35286 to D35389).

The "make index" breakage for LLVM_DEFAULT=14 has been dealt with by
enforcing use of LLVM-13 for www/firefox (in commit 6b2cf7d799d).

This reverts commit febcdcce6160c7203722afca2c226aeac531c1e6.