security/openca-ocspd: New port: OpenCA’s OCSP responder
 - submitter becomes maintainer

OpenCA OCSP Responder is an rfc2560 compliant OCSPD responder.
The server is a stand-alone application and can be integrated into many
different PKI solutions as it does not depend on specific database scheme.
Furthermore it can be used as a responder for multiple CAs.

WWW: https://www.openca.org/projects/ocspd

PR:		264475