games/freeciv: update to 3.0.2 release (+)
WHAT'S CHANGED SINCE 2.6:
The major changes in Freeciv 3.0 are civ2civ3 being the default ruleset,
addition of the alien ruleset, hex maps as the default topology on new installations,
and a whole lot of extra flexibility to create custom rulesets.

As is usual for major releases, 3.0 clients cannot interoperate with pre-3.0 servers,
and vice versa.

Most pre-3.0 savegames can however be loaded into 3.0, and in most cases,
the supplied rulesets have not changed so much as to make it difficult
to complete a game started with 2.6.x's rules.

Release Notes:
	https://freeciv.fandom.com/wiki/NEWS-3.0.2
	https://freeciv.fandom.com/wiki/NEWS-3.0.1
	https://freeciv.fandom.com/wiki/NEWS-3.0.0