graphics/png: fix build with clang 15
Building png with clang 15 results in an error:

    contrib/libtests/pngvalid.c:11662:4: error: call to undeclared function 'feenableexcept'; ISO C99 and later do not support implicit function declarations [-Wimplicit-function-declaration]
       feenableexcept(FE_DIVBYZERO | FE_INVALID | FE_OVERFLOW);
       ^

This is because clang 15 introduced this new warning, and fenv.h only
declares feenableexcept() when compiling for C standards before C11.

The least disruptive way to solve this is to make the port build with
USE_CSTD=gnu89.

PR:		265521
Approved by:	portmgr (antoine)
MFH:		2022Q3