x11-wm/stumpwm: Fix build by updating to 22.05-157
SBCL removed the symbol sb-unix:unix-select in 7882c92e6.  As StumpWM
used this function, it could no longer build when lang/sbcl was updated
to 2.2.7.  This update pulls in a fix which replaces sb-unix:unix-select
with sb-unix:unix-fast-select.

Thanks to Catie in the #stumpwm IRC channel on Libera.Chat for
assistance with debugging and testing.

Sponsored by:	The FreeBSD Foundation