x11/cde-25: Allow cde port/pkg coexistence
Allow cde-25 to coexist with other versions of CDE installed on the
same system.  This allows users to migrate or quickly revert back
should a user wish to temporarily do so. The reason for this is that
CDE 2.5.0 will use fonts such as gb2312.1980 which can mess up fonts
displayed by legacy Xm nd Xaw apps.

In such a scenaroio the user would define CDE_CONCURRENT in their
make.conf or poudriere.d/make.conf, then build and install the port.
When CDE_CONCURRENT is defined, the port is be installed in
${PREFIX}/dt25. To switch between the two CDEs a user could either
manually create a symlink as /usr/local/dt and possibly switch their
~/.dt from one environment or the other. Or one could use something
similar to the following in their .xsession or .xstart file:

	if [ -d .dt-25 ]; then
		rm -f .dt
		ln -s .dt-25 .dt
	fi

The use of this will be expanded as I expect to import a cde-devel
port for those adventurous enough to track the development branch
but who still want the comfort of being able to fall back to a more
stable CDE as needed.

The reason only x11/cde-25 gets this feature is that the install
script provided by cde-24 does not allow for such an enhancement
without significant changes to the port. As cde-24 will be deprecated
at some point, it is not modified. The upcoming cde-devel port will
also include this feature.