deskutils/calendar: update to version 0.9.3
Update the man page to match the version the -CURRENT base system.

The only change is a better description of the // single line comment
syntax, i.e. that // is only considered to start a comment when it
appears at the start of the line or after white space. This allows to
embed URLs in calendar entries.