textproc/python-pptx: new port
python-pptx is a Python library for creating and updating PowerPoint
(.pptx) files.

A typical use would be generating a customized PowerPoint presentation
from database content, downloadable by clicking a link in a web
application. Several developers have used it to automate production
of presentation-ready engineering status reports based on information
held in their work management system. It could also be used for
making bulk updates to a library of presentations or simply to
automate the production of a slide or two that would be tedious to
get right by hand.

PR:	265763