devel/qt5: don't hardcode Perl path
devel/qt5-qmake uses SHEBANG_FILES to update interpreter paths in
multiple files. When built for hybrid ABI, the qt5 patch does not match
shebangs modified for SHEBANG_FILES.

The Perl paths were hardcoded because CheriABI Perl didn't build.
Since Perl was fixed [1], let's not hardcode the paths but use a Perl
binary matching an ABI for which a qt5 port is built.

[1] https://github.com/CTSRD-CHERI/cheribsd-ports/commit/c294835441713deda50fb938c15b456c47c52e96