sysutils/lsof: net/raw_cb.h no longer exists
As of f63cb32c1988561136fabdcc54d16cd200b666d9 (in src) the net/raw_cb.h
header no longer exists. It was removed as part of the retirement of
4.4BSD raw sockets.

Happily this header is not actually required for lsof, so we can just
not include it to fix the ports build.

Reviewed by:	emaste, ler
Sponsored by:	Rubicon Communications, LLC ("Netgate")
Differential Revision:	https://reviews.freebsd.org/D36240