Mk/bsd.options.mk: introduce ABI
ABI is set to a list of ABIs for which ports are being built. At the
moment, the only possible ABI value is "purecap" which is set for
aarch64c and riscv64c architectures.

ABI can be used to mark ports as broken, define options,
activate options by default and exclude options unsupported for a given
ABI.