Uses/python.mk: revert unapproved feature addition (and consumers)
Change proposal was in discussion with open questions and additional
documented design requirement [0] for submitter to review and provide
feedback on, which was not provided.

Submitter (and anyone else) is welcome to work with python@ to
produce an appropriately reviewed feature.

[0] https://wiki.freebsd.org/Python/PEP-517

With hat: python (vishwin, koobs)
PR: 255722, 265660, 265692, 265693
Approved by: fluffy (mentor)