Uses/cargo: Allow using any Rust version if requested
Skip the Rust version check when CARGO_BUILDDEP=any-version

The current version check is justified by USES=cargo making use
of newer toolchain features that not all Cargo versions support
and the fact that Rust binaries are statically linked against
libstd (and others) which have been vulnerable in the past.
We can enforce the use of the correct toolchain only with the
version check. Together with revision bumps of all ports that
have lang/rust as input we can ensure that they are kept "fresh"
and relinked whenever lang/rust is updated.

According to amdmi3@ skipping the check might be useful for build
testing in some cases.

Individual ports should not set CARGO_BUILDDEP=any-version. It
can be set in make.conf or on the command line by users.

PR:		265062
Reported by:	amdmi3