sysutils/nut-devel: Disable new nut_monitor feature
At the moment this new feature, a GUI to monitor nut, inconsistently
builds and installs itself. It fails to build when built under
poudriere or when using an unprivileged account. However it will
build and install when built by hand by root. Even though ./configure
(configure.ac) is supposed to support enabling and disabling the
option, this does not work. The option remains "auto" and nut_monitor
will install or not install based on some unknown at the moment
criteria.