CONFLICTS_INSTALL: check later for conflicts
Until now, the check of the CONFLICTS_INSTALL was done during the sanity
check, meaning very early in the package building process.

this makes it painful in 2 cases:
1. a port depend at build time on the extraction of this port, it will print
   its warning message along with the default timer, while we will never reach
   the said conflict at all.
2. since the ports tree has been staged, one can even go up to create a valid
   package even if a locally conflicting package is locally installed.

This change make the check and print of the conflicts only happen during the
install phase, the same way the ports tree checks if a previous version is
already installed.

PR:	213890
Discussed with:	tcberner
Reviewed by:	tcberner
Differential Revision:	https://reviews.freebsd.org/D36355