java/openjdk11: fix build on 13.1 and -CURRENT
Recent LLVM versions issue warnings for "|" being used with boolean
operands. Such operations exist in OpenJDK and the source lines are
marked with comments to not use logical operations instead.

This breaks the build on 13.1 and -CURRENT when building with LLVM-13
or newer due to the -Werror option being passed to the compiler.

Fix this issue by casting one of the boolean operands to int, as
suggested by the compiler in the error message for the case that the
"|" operation should actually be used on purpose.

I had submitted this patch as an attachment to PR 265687 on
2022-08-07, but it has been ignored by the maintainer (java team).

PR:		265687
Approved by:	portmgr (implicit)