Revert "sysutils/nut*: Add upslog support for multiple UPSs"
upslog is failing to make multiple connections to upsd resulting
in some UPSs logging and others not. This is likely some bug
in upsd's socket handling.

Even though this patch worked for me locally it did not for the
person reporting the problem.

A better approach would be to use select() instead.

PR:	265963

This reverts commit dec4ecc3afecab035a4d4fa59c35093d9f53ffd5.