editors/vim-colorschemes-legacy: Add port
Upstream vim modernized their colorschemes, which is a good thing.
However, the look of them has changed, which is jarring at best for
users used to the original look.

Upstream colorscheme development was split into a new repo, and the
original colorschemes were imported there. This port installs those
colorschemes.

The rest of that repo are the colorschemes included in vim, so there's
not really any benefit to packaging anything else from it.

PR:		265798