devel/lua-language-server: Add port
This was difficult to port largely because of luamake, which:
1) Is itself built by a ninja definition file precompiled just for
   FreeBSD which doesn't work on FreeBSD.
2) Has no documentation whatsoever. I'd love to get rid of the
   BINARY_ALIAS= gcc=cc hack, but I have literally no idea where
   luamake picks up its compiler from.

Any advice on the above is very welcome! sumneko, luamake (and
associated dep bee.lua) went out of their way to try to support FreeBSD,
so it'd be nice to get it working right without heavy patching.

The Sumneko Lua Language Server is a comprehensive Lua development
server. It functions as an LSP client, supporting:

* Over 20 supported annotations for documenting your code
* Go to definition
* Dynamic type checking
* Find references
* Diagnostics/Warnings
* Syntax checking
* Element renaming
* Hover to view details on variables, functions, and more
* Autocompletion
* Support for libraries
* Code formatting
* Spell checking
* Custom plugins

WWW: https://github.com/sumneko/lua-language-server