lang/chicken: restore unsuffixed chicken port
The lang/chicken5 port was introduced to help migrating from v4 in
lang/chicken. The latter is not long gone, and we're left with a '5'
suffix everywhere.  Let's put a copy of chicken v5 in lang/chicken, with
a plan to remove lang/chicken5 in two months.

The use of PORTEPOCH is necessary because it was there in chicken v4.