Mk/bsd.port.mk: do not use master site as default for WWW
Maintainers should set WWW to a website with useful information for
users of the package.

The first MASTER_SITE of the port had been considered to be a useful
default value for the WWW variable.

It has been found not to be a generally useful URL though, therefore
stop providing a default value for WWW.

Approved by:	portmgr (tcberner)