science/mercantile: move URL from pkg-desxcr to Makefile
The URL at the end of the pkg-descr file has been moved into the
Makefile for all existing ports.

BTW: The following URL might be a better value for WWW, since it
provides the user of the package with examples information about
the concepts.

    https://mercantile.readthedocs.io/en/latest/

There is also a link to the GitHub reposirory in that text, but
the WWW value is meant to help a user of the port, not to keep
a note about the repository for the maintainer.

Approved by:	portmgr (implicit)