devel/cmake*: bump to 3.24.0
Update the CMake ports to 3.24. A previous exp-run -- thanks
antoine@ -- showed up new failures. That is normal for a CMake
update. tcberner@ fixed most of them (in the linked PR) and
lang/zig just landed. The original patch for the update came
from tcberner@ as well, but I had to re-jig it after CMake
was reorganized into a metaport and -core package.

PR:		265652