security/cryptopp: Update to 8.7.0
Drop most of the SIMD logic. The cryptopp Makefile autodetects CPU
features regardless if CPUTYPE is set and it is not possible to disable
specific instruction sets. It is only possible to completely disable
assembly optimizations, which is recommended for package building unless
all consumer machines support the build machine's CPU instruction sets.

Since version 8.1, it is no longer necessary to export
-DCRYPTOPP_DISABLE_ASM to the pkgconfig file. Consumers will now
build without it defined if the SIMD option is disabled. Remove old
workarounds pertaining to this issue.
https://github.com/weidai11/cryptopp/issues/779

Bump PORTREVISION on consumers due to ABI incompatibility.

https://www.cryptopp.com/release870.html