shells/ksh2020: Deprecate unmaintained ksh2020
ksh93 modernization development, which was halted in 2020, resulted in
the att/ast team saving the development in their ksh2020 branch and
reverting the mainline back to ksh93. Since the the ksh2020 developers
have forked ksh93 and resumed ksh2020 development under the ksh93
account in their ksh repository. Therefore the now "abandonware" ksh2020
ports is to be removed in favour of the shells/ksh or shells/ksh-devel
ports.

The shells/ksh port tracks releases while shells/ksh-devel tracks
development. shells/ksh93 is available for legacy applications that
require the original AT&T KSH93. The "co-existence" options of each
of these ports allow users to install shells/ksh (and/or ksh-devel) with
shells/ksh93 for users who wish to use the currently developed ksh93 but
still need the legacy AT&T ksh93 due to incompatibilties between the old
and the new. (Note: The latest CDE-25 develpment has switched to using the
new ksh93 shell.)