graphics/cairo: make documentation available in devhelp
GNOME devhelp looks for documentation in particular places,
and cairo was putting it somewhere subtly different.
Switch DOCSDIR (since it was already being used in the
Makefile and pkg-plist, this is the smallest change)
to point to where it should. This approach is copied
from devel/libsoup/ and was suggested by Sean Champ.

PR:		263852 263785