security/putty: permanently remove GSSAPI_DYNAMIC
While it seems a nice idea to load the GSS provider library at run-time,
the existing search code hardwires library versions and only works
with krb5 (MIT Kerberos), and does not attempt to control library paths.

This sort of defeats the purpose, and we do not need a brittle
alternative for a robust GSSAPI_MIT.

So, drop the commented-out GSSAPI_DYNAMIC stuff from the Makefile.