games/openarena: fix version number going backwards
This ports needs to always define the PORTREVISION macro to some value
to prevent it from defaulting to the PORTREVISION set in the ioquake3
port.

Since the ioquake3 port was at PORTREVISION 17, and this setting had
leaked into the openarena and openarena-server ports, setting the
PORTREVISION to 1 in openarena caused the version number to decrease.

Set PORTREVISION to 18 (1 higher than the current value in ioquake3)
to make its version compare higher than before the recent port update.

This commit fixes the version number in both openarena and
openarena-server.