audio/strawberry: Update to 1.0.9
Convert to FLAVORS to allow building against either Qt5 or Qt6 with
the qt5 flavor being the default for now as Qt6 is rather new to
the ports tree.

Remove SUBSONIC and TIDAL options. These are services are now enabled
by default as they don't require additional dependencies.

Several options have been renamed to use standard descriptions and
option descriptions have otherwise been improved.
AUDIOCD -> CDIO
MTPDEV -> MTP

It is now possible to build with the GSTREAMER and/or the VLC
backend, with GSTREAMER remaining the default as several options
require it and it generally provides more features.

NLS is now optional and regression tests have been moved into an
option.

Upstream has moved to using ICU for transliteration of filename
characters, but optional ICONV support is still available.

Improve COMMENT and pkg-descr.

https://github.com/strawberrymusicplayer/strawberry/releases/tag/1.0.8
https://github.com/strawberrymusicplayer/strawberry/releases/tag/1.0.9