Mk/bsd.port.mk: support multiple URLs in WWW
The WWW macro was supposewd to hold only one URL pointing to the most
useful starting point for a user searching for additional information,
configuration examples, etc.

Some port maintainers have set the WWW macro to a list of URLs,
breaking assumptions by scripts and tools in the ports system and of
external users of the ports system.

This change allows WWW to hold a list of URLs by only considering the
first list element in places where only a single URL is supported.

This mirrors the previous behavior of only using the first line tagged
with "WWW: " in the pkg-descr file.

Approved by:	portmgr (tcberner)
Differential Revision:	https://reviews.freebsd.org/D36558