multimedia/libvpx: use CXX as LD
libvpx passes compiler flags to a linker command and links libwebm that
is written in C++. Use a C++ compiler command as a linker command.

This issue is not triggered in environments using a default toolchain
because the ports system doesn't set LD for GNU configure scripts by
default. In case a user sets LD to a custom command or uses a cross
toolchain, libvpx tries to incorrectly use the linker.