sysutils/u-boot-sifive-fu740: Add upstream DTS patch for DA9063 RTC support
This allows the DA9063 RTC driver to probe and attach (and exposes other
MFD children too, but we don't have drivers for them). It also includes
a couple of miscellaneous improvements synced from the Linux DTS,
fixing the compatible string for the on-board QSPI flash and adding the
card detect GPIO for the QSPI SD card.

Reviewed by:	manu
Approved by:	manu (uboot, ports)
Differential Revision:	https://reviews.freebsd.org/D36556