www/caddy: Provide a version string
When build from ports, `caddy version` returns an empty string.

Upstream likes caddy to be built from their xcaddy command, for which
FreeBSD has a port (www/xcaddy) and a package (pkg install xcaddy).
However, our package builders cannot use xcaddy itself, as it needs
external network access. Caddy itself contains no ability to set the
version without the extensive patching that xcaddy performs.

This commit adds a version string to allow the version to be set.
When built with xcaddy, the version string can contain a checksum of
some sort, but we are just using ${DISTVERSIONFULL} here (e.g. v2.6.1).

The lack of a version string has been brought up a number of times to
me, so I'm happy to put this issue to bed.

See https://github.com/caddyserver/caddy/issues/5065 for the latest
discussion about it.